+++ src/App.module.css
@@ -1,10 +1,11 @@
-.card {
-  position: relative;
+.app {
   display: flex;
   flex-direction: column;
-  justify-content: center;
   align-items: center;
-  max-width: 272px;
-  width: 100vw;
+  min-height: 100vh;
+  width: 100%;
+  min-width: 1280px;
+  background-color: #131316;
   box-sizing: border-box;
 }
+
