--- src/components/app/App.module.css
@@ -1,19 +1,0 @@
-.app {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  min-height: 100vh;
-  width: 100%;
-  min-width: 1280px;
-  background-color: #131316;
-  box-sizing: border-box;
-}
-
-.main {
-  display: flex;
-  flex: 1;
-  gap: 40px;
-  max-width: 1240px;
-  padding: 0;
-  box-sizing: border-box;
-}
