+++ src/components/constructor/Constructor.module.css
@@ -1,0 +1,9 @@
+
+.constructor {
+  display: flex;
+  flex: 1;
+  gap: 40px;
+  max-width: 1240px;
+  padding: 0;
+  box-sizing: border-box;
+}
