+++ src/components/header/Header.module.css
@@ -1,0 +1,38 @@
+.header {
+  margin: 0 auto;
+  width: 100%;
+}
+
+.navigation {
+  width: 1280px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin: 0 auto;
+  padding: 0;
+}
+
+.wrapper{
+  composes:navigation;
+  width: 61.5%;
+  margin: 0;
+}
+
+.menuList {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+  display: flex;
+}
+
+.link {
+  display: flex;
+  justify-content: flex-start;
+  color: #8585ad;
+  text-decoration: none;
+}
+
+.linkActive {
+  composes: link;
+  color: #f2f2f3;
+}
