+++ src/components/ingredientDetails/IngredientDetails.jsx
