+++ src/components/ingredientDetails/IngredientDetails.module.css
@@ -12,5 +12,5 @@
   gap: 20px;
   margin: 0;
   padding: 0;
-  list-style:s none;
+  list-style: none;
 }
