+++ src/components/ingredientDetails/components/NutritionValue.jsx
