+++ src/components/ingredientDetails/components/NutritionValue.module.css
