+++ src/constants/index.js
@@ -1,3 +1,4 @@
+// реэкспорт
 export { ariaLable } from './ariaLabel';
 export { productPropType } from './propTypes';
 export { ProductType } from './variables';
