+++ src/utils/order.js
@@ -1,5 +1,3 @@
-// заказ
-// eslint-disable-next-line import/prefer-default-export
 export const order = [
   {
     _id: '60666c42cc7b410027a1a9b1',
